requirements.txt
discord==0.0.2
python-dotenv==0.9.0
praw==6.0.0
